Use 8bit

+++ mlir/tools/mlir-vulkan-runner/VulkanRuntime.h
@@ -115,7 +115,7 @@ public:
   void setResourceData(const DescriptorSetIndex desIndex,
                        const BindingIndex bindIndex,
                        const VulkanHostMemoryBuffer &hostMemBuffer);
-  void setShaderModule(uint32_t *shader);
+  void setShaderModule(uint8_t *shader);
   void setNumWorkGroups(const NumWorkGroups &nWorkGroups);
   void setResourceStorageClassData(const ResourceStorageClassData &stClassData);
   void setEntryPoint(llvm::StringRef entryPointName);
@@ -221,7 +221,7 @@ private:
 
   NumWorkGroups numWorkGroups;
   std::string entryPoint;
-  uint32_t *binary;
+  uint8_t *binary;
   uint32_t binarySize{0};
 
   //===--------------------------------------------------------------------===//
