+++ mlir/tools/mlir-vulkan-runner/vulkan-runtime-wrappers.cpp
@@ -32,7 +32,7 @@ class VulkanRuntimeManager {
       vulkanRuntime.setNumWorkGroups(numWorkGroups);
     }
 
-    void setShaderModule(uint32_t *shader) {
+    void setShaderModule(uint8_t *shader) {
       vulkanRuntime.setShaderModule(shader);
     }
 
@@ -74,7 +74,7 @@ void setNumWorkGroups(uint32_t x, uint32_t y, uint32_t z) {
   VulkanRuntimeManager::instance()->setNumWorkGroups({x, y, z});
 }
 
-void setBinaryShader(uint32_t *shader) {
+void setBinaryShader(uint8_t *shader) {
   VulkanRuntimeManager::instance()->setShaderModule(shader);
 }
